Malloc fixes (#70)
* Several memory allocation fixes, primarily motivated by the fact that the Diffie-Hellman implementation in mbedtls is sensitive to misaligned allocations.  All malloc's are now 16-byte aligned.  This PR has several other points where remalloc was used instead of realloc, or memory needed to be zeroed upon allocation.  Finally, this PR also standardizes the definition of assert across layers, so that code in the lib directory can both use assertions and link properly in the PAL and shim.

+++ Pal/lib/api.h
@@ -111,7 +111,7 @@ struct config_store {
     struct list_head root, entries;
     void *           raw_data;
     int              raw_size;
-    void *           (*malloc) (int);
+    void *           (*malloc) (size_t);
     void             (*free) (void *);
 };
